+++ edgy/cli/operations/shell/ptpython.py
@@ -28,23 +28,25 @@ def get_ptpython(app: typing.Any, registry: Registry, options: typing.Any = None
         from ptpython.repl import embed, run_config
 
         def run_ptpython() -> None:
-            imported_objects = import_objects(app, registry)
             history_filename = os.path.expanduser("~/.ptpython_history")
 
             config_file = os.path.expanduser(settings.ptpython_config_file)
-            if not os.path.exists(config_file):
-                embed(
-                    globals=imported_objects,
-                    history_filename=history_filename,
-                    vi_mode=vi_mode(),
-                )
-            else:
-                embed(
-                    globals=imported_objects,
-                    history_filename=history_filename,
-                    vi_mode=vi_mode(),
-                    configure=run_config,
-                )
+            with registry.with_async_env():
+                # we need an initialized registry first to detect reflected models
+                imported_objects = import_objects(app, registry)
+                if not os.path.exists(config_file):
+                    embed(
+                        globals=imported_objects,
+                        history_filename=history_filename,
+                        vi_mode=vi_mode(),
+                    )
+                else:
+                    embed(
+                        globals=imported_objects,
+                        history_filename=history_filename,
+                        vi_mode=vi_mode(),
+                        configure=run_config,
+                    )
 
     except (ModuleNotFoundError, ImportError):
         error = "You must have ptpython installed to run this. Run `pip install ptpython`"